refactor run cmd for py/ts
