+++ apps/framework-cli/src/framework/scripts.rs
@@ -4,8 +4,8 @@ use super::languages::SupportedLanguages;
 
 pub mod collector;
 pub mod config;
-pub mod errors;
 pub mod executor;
+pub mod utils;
 
 use crate::framework::python::templates::PYTHON_BASE_SCRIPT_TEMPLATE;
 use crate::framework::scripts::config::WorkflowConfig;
