+++ apps/framework-cli/src/framework/typescript.rs
@@ -2,7 +2,6 @@ pub mod bin;
 pub mod blocks;
 pub mod checker;
 pub mod consumption;
-pub mod executor;
 pub mod export_collectors;
 pub mod generator;
 pub mod parser;
